requirements.txt
pandas>=1.5.0
requests>=2.28.0
matplotlib>=3.6.0
plotly>=5.11.0
python-dotenv>=0.19.0
icalendar>=5.0.0
PyYAML>=6.0
